change id to Integer, implement UserDetails

+++ src/main/java/com/example/ProductsTracker/user/UserController.java
@@ -37,17 +37,17 @@ public class UserController {
     }
 
     @GetMapping("{id}")
-    public ResponseEntity<User> getUserById(@PathVariable("id") Long userId) {
+    public ResponseEntity<User> getUserById(@PathVariable("id") Integer userId) {
         return new ResponseEntity<User>(userService.getUserById(userId), HttpStatus.OK);
     }
 
     @PatchMapping("{id}")
-    public ResponseEntity<User> updateUser(@RequestBody User user, @PathVariable("id") Long userId) {
+    public ResponseEntity<User> updateUser(@RequestBody User user, @PathVariable("id") Integer userId) {
         return new ResponseEntity<User>(userService.patchUser(user, userId), HttpStatus.CREATED);
     }
 
     @DeleteMapping("{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable("id") Long userId){
+    public ResponseEntity<String> deleteUser(@PathVariable("id") Integer userId){
         userService.deleteUser(userId);
         return new ResponseEntity<>("User deleted successfully", HttpStatus.OK);
     }
